-Codechange: rename HQX common.h to hqx_common.h
