+++ projects/opendune_vs100.vcxproj
@@ -429,7 +429,6 @@
     <ClCompile Include="..\src\table\windowdesc.c">
       <ObjectFileName>$(IntDir)src\table\</ObjectFileName>
     </ClCompile>
-    <ClInclude Include="..\src\video\common.h"/>
     <ClCompile Include="..\src\video\hq2x.c">
       <ObjectFileName>$(IntDir)src\video\</ObjectFileName>
     </ClCompile>
@@ -444,6 +443,7 @@
       <ObjectFileName>$(IntDir)src\video\</ObjectFileName>
     </ClCompile>
     <ClInclude Include="..\src\video\hqx.h"/>
+    <ClInclude Include="..\src\video\hqx_common.h"/>
     <ClCompile Include="..\src\video\hqx_init.c">
       <ObjectFileName>$(IntDir)src\video\</ObjectFileName>
     </ClCompile>
