+++ src/video/hqx_common.h
